+++ Lab5/EquilateralTriangleLawn.h
@@ -16,6 +16,6 @@ namespace lab5
 		virtual unsigned int GetFencePrice(eFenceType fenceType) const override;
 
 	private:
-		const unsigned int mWidth;
+		const double mWidth;
 	};
 }
